requirements.txt
absl-py==0.15.0
anyio==3.3.4
argon2-cffi==21.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
contextvars==2.4
cycler==0.11.0
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
easydict==1.9
entrypoints==0.3
fire==0.4.0
Flask==2.0.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.1.0
idna==3.3
imageio==2.10.4
imageio-ffmpeg==0.4.5
immutables==0.16
importlib-metadata==4.8.1
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyter-server==1.11.2
jupyterlab==3.2.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mediapipe==0.8.3
mistune==0.8.4
moviepy==1.0.3
mysql-connector-python==8.0.27
nbclassic==0.3.4
nbclient==0.5.5
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.19.3
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==21.2
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pgi==0.0.11.2
pickleshare==0.7.5
Pillow==8.4.0
proglog==0.1.9
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.21
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.15.0
sniffio==1.2.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thop==0.0.31.post2005241907
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
wrapt==1.12.1
yolov5==6.0.4
zipp==3.6.0
